Revert "Chroma Issue 536: Broke down index.ts into multiple files"
This reverts commit a58acdf74f19f23b2518bdd5287eb522cdb46e49.
