+++ clients/js/examples/browser/app.ts
@@ -1,5 +1,5 @@
 // import env.ts
-import { ChromaClient } from "../../src/ChromaClient";
+import { ChromaClient } from "../../src/index";
 
 window.onload = async () => {
   const chroma = new ChromaClient("http://localhost:8000");
