--- clients/js/src/embeddings/IEmbeddingFunction.ts
@@ -1,3 +1,0 @@
-export interface IEmbeddingFunction {
-    generate(texts: string[]): Promise<number[][]>;
-}
